Feature/prac

+++ src/main/java/com/example/trip/service/FeedService.java
@@ -1,5 +1,6 @@
 package com.example.trip.service;
 
+import com.example.trip.advice.exception.*;
 import com.example.trip.domain.*;
 import com.example.trip.dto.request.FeedRequestDto;
 import com.example.trip.repository.LikeRepository;
@@ -24,14 +25,15 @@ public class FeedService {
     private final LikeRepository likeRepository;
     private final BookMarkRepository bookMarkRepository;
     private final CommentRepository commentRepository;
-private final AwsS3Service awsS3Service;
-    public List<Feed> findAll(){
+    private final AwsS3Service awsS3Service;
+
+    public List<Feed> findAll() {
 
         return feedRepository.findAll();
     }
 
     @Transactional
-    public List<Long> registerFeed(User user, FeedRequestDto.FeedRequestRegisterDto feedRequestRegisterDto){
+    public List<Long> registerFeed(User user, FeedRequestDto.FeedRequestRegisterDto feedRequestRegisterDto) {
         // feed 저장
         Feed feed = Feed.builder()
                 .user(user)
@@ -51,7 +53,7 @@ private final AwsS3Service awsS3Service;
 
         List<List<FeedDetailLoc>> feedDetailLocs = feedDetails.stream()
                 .map(FeedDetail::getFeedDetailLoc)
-                        .collect(Collectors.toList());
+                .collect(Collectors.toList());
 
         feedDetailLocs.stream()
                 .forEach(x -> feedDetailLocRepository.saveAll(x));
@@ -73,9 +75,11 @@ private final AwsS3Service awsS3Service;
         return newFeedDetailLocs;
     }
 
-    public Map<String, String> registerFeedImage(Long feedDetailLocId, List<MultipartFile> imgFiles){
-        Map<String, String> nameAndUrl  = awsS3Service.uploadFile(imgFiles);
-        FeedDetailLoc feedDetailLoc = feedDetailLocRepository.findById(feedDetailLocId).orElseThrow(() -> new NullPointerException(""));
+    public Map<String, String> registerFeedImage(Long feedDetailLocId, List<MultipartFile> imgFiles) {
+        Map<String, String> nameAndUrl = awsS3Service.uploadFile(imgFiles);
+        // 해당 피드 상세 위치 값이 있는지 체크
+        FeedDetailLoc feedDetailLoc = feedDetailLocRepository.findById(feedDetailLocId)
+                .orElseThrow(() -> new FeedDetailLocNotFoundException());
         nameAndUrl.entrySet().forEach(x -> feedDetailLocImgRepository.save(FeedDetailLocImg.builder()
                 .feedDetailLoc(feedDetailLoc)
                 .fileName(x.getKey())
@@ -85,14 +89,22 @@ private final AwsS3Service awsS3Service;
         return nameAndUrl;
     }
 
-    public void deleteFeedImage(FeedRequestDto.FeedRequestDeleteImgDto  FeedRequestDeleteImgDto){
-        List<String> fileNames  = FeedRequestDeleteImgDto.getFileNames();
+    public void deleteFeedImage(FeedRequestDto.FeedRequestDeleteImgDto FeedRequestDeleteImgDto) {
+        List<String> fileNames = FeedRequestDeleteImgDto.getFileNames();
         fileNames.forEach(x -> awsS3Service.deleteFile(x));
     }
 
     @Transactional
-    public void modifyFeed(Long feedId, FeedRequestDto.FeedRequestModifyDto feedRequestModifyDto){
-        Feed feed = feedRepository.findById(feedId).orElseThrow(() -> new IllegalArgumentException("ID가 존재하지 않습니다. "));
+    public void modifyFeed(User user, Long feedId, FeedRequestDto.FeedRequestModifyDto feedRequestModifyDto) {
+        // 피드를 올린 사람만 권한이 있어야함
+        List<Feed> myFeed = feedRepository.findByIdAndUserId(feedId, user.getId());
+
+        if (myFeed.isEmpty()) {
+            throw new AuthFeedNotFoundException();
+        }
+
+
+        Feed feed = feedRepository.findById(feedId).orElseThrow(() -> new FeedNotFoundException());
 
         //feed Detail 수정
         feedRequestModifyDto.getFeedDetail().stream()
@@ -109,12 +121,27 @@ private final AwsS3Service awsS3Service;
         feed.update(feedRequestModifyDto);
     }
 
-    public void deleteFeed(Long feedId){
+    public void deleteFeed(User user, Long feedId) {
+        // 피드를 올린 사람만 권한이 있어야함
+        List<Feed> myFeed = feedRepository.findByIdAndUserId(feedId, user.getId());
+
+        if (myFeed.isEmpty()) {
+            throw new AuthFeedNotFoundException();
+        }
         feedRepository.deleteById(feedId);
     }
 
-    public void likeFeed(Long feedDetailLocId, User user){
-        FeedDetailLoc feedDetailLoc = feedDetailLocRepository.findById(feedDetailLocId).orElseThrow(() -> new NullPointerException("해당 값이 없습니다."));
+    public void likeFeed(Long feedDetailLocId, User user) {
+        // 해당 피드 상세 위치 값이 있는지 체크
+        FeedDetailLoc feedDetailLoc = feedDetailLocRepository.findById(feedDetailLocId)
+                .orElseThrow(() -> new FeedDetailLocNotFoundException());
+
+        //이미 좋아요한 게시글인지 체크
+        List<Likes> didYouLike = likeRepository.findByFeedDetailLocIdAndUserId(feedDetailLocId, user.getId());
+        if (!didYouLike.isEmpty()){
+            throw new AlreadyLikeException();
+        }
+
         Likes like = Likes.builder()
                 .feedDetailLoc(feedDetailLoc)
                 .user(user)
@@ -124,12 +151,20 @@ private final AwsS3Service awsS3Service;
     }
 
 
-    public void unlikeFeed(Long feedId, User user){
-        likeRepository.deleteLikeFeed(feedId, user.getId());
+    public void unlikeFeed(Long feedDetailLocId, User user) {
+        // 좋아요를 한 사람만 권한이 있어야함
+        List<Likes> myLike = likeRepository.findByFeedDetailLocIdAndUserId(feedDetailLocId, user.getId());
+
+        if (myLike.isEmpty()) {
+            throw new AuthLikesNotFoundException();
+        }
+        likeRepository.deleteLikeFeed(feedDetailLocId, user.getId());
     }
 
-    public void bookmarkFeed(Long feedDetailLocId, User user){
-        FeedDetailLoc feedDetailLoc = feedDetailLocRepository.findById(feedDetailLocId).orElseThrow(() -> new NullPointerException("해당 값이 없습니다."));
+    public void bookmarkFeed(Long feedDetailLocId, User user) {
+        // 해당 피드 상세 위치 값이 있는지 체크
+        FeedDetailLoc feedDetailLoc = feedDetailLocRepository.findById(feedDetailLocId)
+                .orElseThrow(() -> new FeedDetailLocNotFoundException());
         BookMark bookmark = BookMark.builder()
                 .feedDetailLoc(feedDetailLoc)
                 .user(user)
@@ -137,12 +172,21 @@ private final AwsS3Service awsS3Service;
 
         bookMarkRepository.save(bookmark);
     }
-    public void unbookmarkFeed(Long feedDetailLocId, User user){
+
+    public void unbookmarkFeed(Long feedDetailLocId, User user) {
+        // 북마크를 한 사람만 권한이 있어야함
+        List<BookMark> myBookmark = bookMarkRepository.findByFeedDetailLocIdAndUserId(feedDetailLocId, user.getId());
+
+        if (myBookmark.isEmpty()) {
+            throw new AuthBookMarkNotFoundException();
+        }
         bookMarkRepository.deleteBookmarkFeed(feedDetailLocId, user.getId());
     }
 
-    public void registerFeedComment(User user, Long feedDetailLocId, FeedRequestDto.FeedRequestCommentRegisterDto feedRequestCommentRegisterDto){
-        FeedDetailLoc feedDetailLoc = feedDetailLocRepository.findById(feedDetailLocId).orElseThrow(() -> new NullPointerException("해당 값이 없습니다."));
+    public void registerFeedComment(User user, Long feedDetailLocId, FeedRequestDto.FeedRequestCommentRegisterDto feedRequestCommentRegisterDto) {
+        // 해당 피드 상세 위치 값이 있는지 체크
+        FeedDetailLoc feedDetailLoc = feedDetailLocRepository.findById(feedDetailLocId)
+                .orElseThrow(() -> new FeedDetailLocNotFoundException());
         FeedComment feedComment = FeedComment.builder()
                 .feedDetailLoc(feedDetailLoc)
                 .user(user)
@@ -153,13 +197,26 @@ private final AwsS3Service awsS3Service;
     }
 
     @Transactional
-    public void modifyFeedComment(Long commentId, FeedRequestDto.FeedRequestCommentModifyDto feedRequestCommentModifyDto){
+    public void modifyFeedComment(User user, Long commentId, FeedRequestDto.FeedRequestCommentModifyDto feedRequestCommentModifyDto) {
+        // 댓글을 올린 사람만 권한이 있어야함
+        List<FeedComment> myFeedComment = commentRepository.findByIdAndUserId(commentId, user.getId());
+
+        if (myFeedComment.isEmpty()) {
+            throw new AuthFeedCommentNotFoundException();
+        }
         FeedComment feedComment = commentRepository.findById(commentId).orElseThrow(() -> new NullPointerException("해당 값이 없습니다."));
         feedComment.update(feedRequestCommentModifyDto);
 
     }
 
-    public void deleteFeedComment(Long commentId){
+    public void deleteFeedComment(User user, Long commentId) {
+        // 댓글을 올린 사람만 권한이 있어야함
+        List<FeedComment> myFeedComment = commentRepository.findByIdAndUserId(commentId, user.getId());
+
+        if (myFeedComment.isEmpty()) {
+            throw new AuthFeedCommentNotFoundException();
+        }
+
         commentRepository.deleteById(commentId);
     }
 }
